Remove RunMetrics; add RunContext.next_step
Remove the internal RunMetrics implementation and related Metric classes, centralizing global step handling on RunContext via a new next_step method. Deleted swanlab/sdk/internal/context/metrics.py and updated imports/usages across the SDK: RunContext no longer exposes a metrics attribute, Run.log now calls ctx.next_step, the monitor no longer advances system steps, and consumers/builders no longer maintain or define metric state locally (RecordBuilder.build_column_from_log removed; build_column_from_scalar_define now returns only a ColumnRecord). Also adjusted debug messages in core_python. This streamlines metric handling and moves step management into the RunContext.
