+++ swanlab/sdk/internal/run/components/builder/__init__.py
@@ -6,7 +6,7 @@
 """
 
 from functools import singledispatchmethod
-from typing import Optional, Tuple, Type, Union
+from typing import Optional
 
 from google.protobuf.timestamp_pb2 import Timestamp
 
@@ -21,8 +21,7 @@ from swanlab.proto.swanlab.metric.column.v1.column_pb2 import (
 from swanlab.proto.swanlab.metric.data.v1.data_pb2 import MediaRecord
 from swanlab.proto.swanlab.system.v1.console_pb2 import ConsoleRecord
 from swanlab.sdk.internal.bus.events import ConfigEvent, ConsoleEvent, ParseResult, ScalarDefineEvent
-from swanlab.sdk.internal.context import RunContext, TransformData, TransformMedia
-from swanlab.sdk.internal.context.metrics import MediaMetric, ScalarMetric
+from swanlab.sdk.internal.context import RunContext, TransformMedia
 from swanlab.sdk.internal.pkg import adapter, console, fs
 from swanlab.sdk.internal.run.transforms import ECharts, Scalar, echarts
 
@@ -124,36 +123,9 @@ class RecordBuilder:
         )
         return media_record, cls
 
-    def build_column_from_log(
-        self,
-        cls: Type[TransformData],
-        key: str,
-    ) -> Tuple[ColumnRecord, Union[ScalarMetric, MediaMetric]]:
-        """隐式创建列：从 TransformType 推断 ColumnType，并同步 RunMetrics"""
-        column_record = ColumnRecord(
-            column_key=key,
-            column_type=cls.column_type(),
-            column_class=ColumnClass.COLUMN_CLASS_CUSTOM,
-            # 自动创建的section默认为公共section
-            section_type=SectionType.SECTION_TYPE_PUBLIC,
-        )
-        col_type = column_record.column_type
-        metrics = self._ctx.metrics
-        metric: Union[ScalarMetric, MediaMetric]
-        if issubclass(cls, TransformMedia):
-            media_type_str = adapter.medium[col_type]
-            metric = metrics.define_media(key=key, column=column_record, path=self._ctx.media_dir / media_type_str)
-        else:
-            metric = metrics.define_scalar(key=key, column=column_record)
-
-        return column_record, metric
-
-    def build_column_from_scalar_define(
-        self,
-        event: ScalarDefineEvent,
-    ) -> Tuple[ColumnRecord, ScalarMetric]:
+    @staticmethod
+    def build_column_from_scalar_define(event: ScalarDefineEvent) -> ColumnRecord:
         """显式创建标量列（DefineEvent）"""
-        metrics = self._ctx.metrics
         section_type = SectionType.SECTION_TYPE_SYSTEM if event.system else SectionType.SECTION_TYPE_PUBLIC
         col = ColumnRecord(
             column_key=event.key,
@@ -166,8 +138,7 @@ class RecordBuilder:
             metric_name=event.name or "",
             metric_colors=MetricColors(light=event.color, dark=event.color) if event.color else None,
         )
-        metric = metrics.define_scalar(key=event.key, column=col)
-        return col, metric
+        return col
 
     # ── 系统元数据 ──
     @staticmethod
